@@ -2,7 +2,7 @@
   "name": "@firestitch/package",
   "title": "package.json -> Rename Title to Component Title",
   "description": "@firestitch/package",
-  "version": "7.0.0",
+  "version": "9.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Firestitch/ngx-package"
@@ -25,44 +25,47 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.9",
-    "@angular-devkit/build-ng-packagr": "^0.803.9",
-    "@angular/animations": "^8.2.10",
-    "@angular/cdk": "^8.2.3",
-    "@angular/cli": "^8.3.9",
-    "@angular/common": "^8.2.10",
-    "@angular/compiler": "^8.2.10",
-    "@angular/compiler-cli": "^8.2.10",
-    "@angular/core": "^8.2.10",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.10",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.10",
-    "@angular/platform-browser-dynamic": "^8.2.10",
-    "@angular/router": "^8.2.10",
-    "@compodoc/compodoc": "^1.1.10",
-    "@firestitch/drawer": "^8.0.0",
+    "@angular-devkit/build-angular": "~0.900.7",
+    "@angular-devkit/build-ng-packagr": "~0.900.7",
+    "@angular/animations": "^9.0.7",
+    "@angular/cdk": "^9.1.3",
+    "@angular/cli": "^9.0.7",
+    "@angular/common": "^9.0.7",
+    "@angular/compiler": "^9.0.7",
+    "@angular/compiler-cli": "^9.0.7",
+    "@angular/core": "^9.0.7",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "^9.0.7",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^9.1.3",
+    "@angular/platform-browser": "^9.0.7",
+    "@angular/platform-browser-dynamic": "^9.0.7",
+    "@angular/router": "^9.0.7",
+    "@compodoc/compodoc": "^1.1.11",
+    "@firestitch/drawer": "^8.2.0",
     "@firestitch/example": "^8.0.0",
     "@firestitch/iframe": "^8.0.0",
-    "@firestitch/menu": "^7.0.11",
-    "@firestitch/message": "^8.0.1",
-    "codelyzer": "^5.1.2",
+    "@firestitch/label": "^8.1.0",
+    "@firestitch/menu": "^9.0.1",
+    "@firestitch/message": "^8.1.0",
+    "codelyzer": "^5.2.1",
     "cpx": "^1.5.0",
     "hammerjs": "^2.0.8",
-    "highlight.js": "^9.15.10",
+    "highlight.js": "^9.18.1",
     "lodash-es": "^4.17.15",
-    "ng-packagr": "^5.6.1",
-    "ngx-toastr": "^11.1.0",
+    "ng-packagr": "^9.0.0",
+    "ngx-toastr": "^12.0.0",
     "npm-run-all": "^4.1.5",
-    "rxjs": "^6.5.3",
-    "tsickle": "^0.37.0",
-    "tslint": "^5.20.0",
-    "typescript": "^3.5.3",
-    "zone.js": "^0.10.2"
+    "rxjs": "^6.5.4",
+    "tslint": "^6.1.0",
+    "typescript": "^3.7.5",
+    "zone.js": "~0.10.3"
   },
   "engines": {
     "node": ">=6.0.0"
   },
-  "peerDependencies": {}
+  "peerDependencies": {},
+  "dependencies": {
+    "tslib": "^1.10.0"
+  }
 }
